+++ docs/0-0-symfony.md
@@ -2,12 +2,12 @@
 
 ## Replacing the default repository type
 
-Replacing Doctrine's default repository type with `Happyr\DoctrineSpecification\EntitySpecificationRepository` is easy
+Replacing Doctrine's default repository type with `Happyr\DoctrineSpecification\Repository\EntitySpecificationRepository` is easy
 in Symfony. The doctrine bundle provides a place in configuration to specify the new type.
 
 ```yml
 // app/config/config.yml
 doctrine:
     orm:
-        default_repository_class: Happyr\DoctrineSpecification\EntitySpecificationRepository
+        default_repository_class: Happyr\DoctrineSpecification\Repository\EntitySpecificationRepository
 ```
