+++ docs/0-1-zend1.md
@@ -5,9 +5,9 @@
 Doctrine integration with Zend Framework 1 is done manually. However, replacing the default repository type is still
 simple. In the bootstrap (or wherever Doctrine is configured in the system in question), use Doctrine configurations
 `setRepositoryFactory` method and provide an implementation of `Doctrine\ORM\Repository\RepositoryFactory`. A basic
-implementation is provided with `Happyr\DoctrineSpecification\RepositoryFactory`.
+implementation is provided with `Happyr\DoctrineSpecification\Repository\RepositoryFactory`.
 
 ```php
 // During doctrine configuration
-$config->setRepositoryFactory(new \Happyr\DoctrineSpecification\RepositoryFactory());
+$config->setRepositoryFactory(new \Happyr\DoctrineSpecification\Repository\RepositoryFactory());
 ```
